refactored the script is modular and it is split in files to improve readability
