+++ index.html
@@ -10,10 +10,10 @@
     <title>Oxygen Shop</title>
   </head>
   <body>
-    <nav class="nav">
+    <nav id='nav' class="nav">
       <article class="nav__oxygenshop">
         <h1 class="nav__oxygenshop__title">OXYGENShop</h1>
-        <article class="nav__oxygenshop__section-list">
+        <article id='nav-oxygenshop-section-list' class="nav__oxygenshop__section-list">
           <div class="nav__oxygenshop__section-list__option" data-value="why-us">WHY US</div>
           <div class="nav__oxygenshop__section-list__option" data-value="benefits">BENEFITS</div>
           <div class="nav__oxygenshop__section-list__option" data-value="prices">PRICES</div>
@@ -21,7 +21,7 @@
         </article>
       </article>
       <article class="nav__menu">
-        <img class="nav__menu__img" src="./public/svg/menu.svg" alt="menu icon"><img/>
+        <img id='nav-menu-img' class="nav__menu__img" src="./public/svg/menu.svg" alt="menu icon"><img/>
       </article>
     </nav>
     <article class="percentage-scroller"></article>
@@ -44,7 +44,7 @@
       </div>
     </article>
     <main>
-      <header>
+      <header id="why-us">
         <article class="sellmore">
           <h1 class="sellmore__title">Sell more!</h1>
           <p class="sellmore__text">Open shop on our platform and increase your sales</p>
@@ -54,7 +54,7 @@
           <img class="split-container__img" src="./public/img/ordenador_background.png" alt="Ordenador fondo" />
         </article>
       </header>
-      <section class="whyus" id="why-us">
+      <section class="whyus" >
         <section class="whyus__mostpopular">
           <h1 class="whyus__mostpopular__title">The most popular sales platform in country</h1>
           <section class="whyus__mostpopular__content">
