+++ js_script/index.js
@@ -1,16 +1,5 @@
-import setupCurrencyExchangeElements from "./setupCurrencyExchangeElements.js";
-import setupFormElements from "./setupFormElements.js";
-import setupMenuElements from "./setupMenuElements.js";
-import setupScrollElements from "./setupScrollElements.js"
-import sliderClass from "./sliderClass.js"
-
-//create Slider
-new sliderClass('slider');
-
-setupScrollElements();
-
-setupMenuElements();
-
-setupFormElements();
-
-setupCurrencyExchangeElements();
+import './imageSlider.js';
+import './form.js';
+import './scroll.js';
+import './currencyExchange.js';
+import './menu.js';
